[tests]: extract test helpers to a separate crate
