+++ test-utils/src/lib.rs
@@ -1,0 +1,4 @@
+//! Shared test helpers for JSONRPC v2.
+
+pub mod helpers;
+pub mod types;
